UISlider和UIScrollbar的OnValueChange事件存在预制体里的是float，但是UE5中改为了double，这样加载的预制体的类型就错了，所以在UE5里强制改为double

+++ Source/LGUI/Public/PlayTween/LGUIPlayTween_Params.h
@@ -17,11 +17,11 @@ protected:
 		float to = 1.0f;
 	/** parameter float is interpolated value from->to */
 	UPROPERTY(EditAnywhere, Category = "Event")
-		FLGUIEventDelegate onUpdateValue = FLGUIEventDelegate(LGUIEventDelegateParameterType::Double);
+		FLGUIEventDelegate onUpdateValue = FLGUIEventDelegate(LGUIEventDelegateParameterType::Float);
 
 	virtual void OnUpdate(float progress)override
 	{
-		onUpdateValue.FireEvent((double)FMath::Lerp(from, to, progress));
+		onUpdateValue.FireEvent(FMath::Lerp(from, to, progress));
 	}
 };
